[refactor] 폴더 엔트리포인트들 Index로 이름변경
