+++ frontend/Idle/src/components/trimDetailModal/index.jsx
